Creando el eliminar y el editar prioridades
